Updated names in AzureData to match MicrosoftSentinel Moved list_sentinel_workspaces to AzureData Added connection checks to Sentinel features
