+++ msticpy/data/context/azure/azure_data.py
@@ -171,7 +171,7 @@ class AzureData:
             raise CloudError("Could not create a Subscription client.")
         self.connected = True
 
-    def get_subscriptions(self) -> pd.DataFrame:
+    def list_subscriptions(self) -> pd.DataFrame:
         """
         Get details of all subscriptions within the tenant.
 
@@ -257,7 +257,7 @@ class AzureData:
             "Spending Limit": sub.subscription_policies.spending_limit,
         }
 
-    def get_resources(  # noqa: MC0001
+    def list_resources(  # noqa: MC0001
         self, sub_id: str, rgroup: str = None, get_props: bool = False
     ) -> pd.DataFrame:
         """
@@ -833,6 +833,50 @@ class AzureData:
                     credential_scopes=[self.az_cloud_config.token_uri],
                 ),
             )
+
+    def list_sentinel_workspaces(self, sub_id: str = None) -> Dict[str, str]:
+        """
+        Return a list of Microsoft Sentinel workspaces in a Subscription.
+
+        Parameters
+        ----------
+        sub_id : str
+            The subscription ID to get a list of workspaces from.
+
+        Returns
+        -------
+        Dict
+            A dictionary of workspace names and ids
+
+        """
+        print("Finding Microsoft Sentinel Workspaces...")
+        res = self.get_resources(sub_id=sub_id)  # type: ignore
+        # handle no results
+        if isinstance(res, pd.DataFrame) and not res.empty:
+            sentinel = res[
+                (res["resource_type"] == "Microsoft.OperationsManagement/solutions")
+                & (res["name"].str.startswith("SecurityInsights"))
+            ]
+            workspaces = []
+            for wrkspace in sentinel["resource_id"]:
+                res_details = self.get_resource_details(
+                    sub_id=sub_id, resource_id=wrkspace  # type: ignore
+                )
+                workspaces.append(res_details["properties"]["workspaceResourceId"])
+
+            workspaces_dict = {}
+            for wrkspace in workspaces:
+                name = wrkspace.split("/")[-1]
+                workspaces_dict[name] = wrkspace
+            return workspaces_dict
+
+        print(f"No Microsoft Sentinel workspaces in {sub_id}")
+        return {}
+
+    # Get > List Aliases
+    get_subscriptions = list_subscriptions
+    get_resources = list_resources
+    get_sentinel_workspaces = list_sentinel_workspaces
 
 
 def get_api_headers(token: str) -> Dict:
