+++ msticpy/data/context/azure/sentinel_utils.py
@@ -39,10 +39,7 @@ class SentinelUtilsMixin:
 
     def _get_items(self, url: str, params: str = "2020-01-01") -> httpx.Response:
         """Get items from the API."""
-        if not self.connected:  # type: ignore
-            raise MsticpyAzureConnectionError(
-                "Ensure you run .connect before calling other functions."
-            )
+        self.check_connected()
         return httpx.get(
             url,
             headers=get_api_headers(self.token),  # type: ignore
@@ -185,6 +182,13 @@ class SentinelUtilsMixin:
                 f"/{res_info['workspace_name']}",
             ]
         )
+
+    def check_connected(self):
+        """Check that Sentinel workspace is connected."""
+        if not self.connected:
+            raise MsticpyAzureConnectionError(
+                "Not connected to Sentinel, ensure you run `.connect` before calling functions."
+            )
 
 
 def _azs_api_result_to_df(response: httpx.Response) -> pd.DataFrame:
