+++ msticpy/data/context/azure/sentinel_watchlists.py
@@ -83,6 +83,7 @@ class SentinelWatchlistsMixin:
             If there is an issue creating the watchlist.
 
         """
+        self.check_connected()
         if self._check_watchlist_exists(watchlist_name):
             raise MsticpyUserError(f"Watchlist {watchlist_name} already exist.")
         watchlist_url = self.sent_urls["watchlists"] + f"/{watchlist_name}"  # type: ignore
@@ -169,6 +170,7 @@ class SentinelWatchlistsMixin:
             If the API returns an error.
 
         """
+        self.check_connected()
         # Check requested watchlist actually exists
         if not self._check_watchlist_exists(watchlist_name):
             raise MsticpyUserError(f"Watchlist {watchlist_name} does not exist.")
@@ -243,6 +245,7 @@ class SentinelWatchlistsMixin:
             If the API returns an error.
 
         """
+        self.check_connected()
         # Check requested watchlist actually exists
         if not self._check_watchlist_exists(watchlist_name):
             raise MsticpyUserError(f"Watchlist {watchlist_name} does not exist.")
